bindings/lua: rename flux-lua dir to flux
Simplify inclusion of flux-specific lua modules by renaming flux-lua
directory to simply flux. This doesn't conflict with flux.so main
flux lua module.
