+++ flux-core.spec.in
@@ -54,7 +54,7 @@ rm -rf $RPM_BUILD_ROOT
 %{_libdir}/flux/connectors/*.so
 %{_libdir}/lua/*/*.so
 %{_prefix}/share/lua/*/*.lua
-%{_prefix}/share/lua/*/flux-lua/*.lua
+%{_prefix}/share/lua/*/flux/*.lua
 /etc/ld.so.conf.d/*
 %{_libexecdir}/flux
 %{_bindir}/flux
