+++ src/cmd/flux-topo
@@ -4,7 +4,7 @@
 
 local flux = require 'flux'
 local hostlist = require ("hostlist")
-local getopt = require 'flux-lua.alt_getopt'.get_opts
+local getopt = require 'flux.alt_getopt'.get_opts
 
 local prog = string.match (arg[0], "([^/]+)$")
